[3.0 Triple] Fix no error signal when channel closed (#9858)
* Fix no error signal when channel closed

* Some reactor

* Fix ut

* Do not send reset when received client rst

* Add more comments

* Do not ignore frame to avoid race condition

* Try fix ut

* Skip assert for debug

* Add service

* Remove context before use

* Fix comment
