+++ dubbo-rpc/dubbo-rpc-triple/src/main/java/org/apache/dubbo/rpc/protocol/tri/call/ClientCall.java
@@ -17,254 +17,97 @@
 
 package org.apache.dubbo.rpc.protocol.tri.call;
 
-import org.apache.dubbo.common.logger.Logger;
-import org.apache.dubbo.common.logger.LoggerFactory;
 import org.apache.dubbo.common.stream.StreamObserver;
-import org.apache.dubbo.remoting.api.Connection;
 import org.apache.dubbo.rpc.TriRpcStatus;
-import org.apache.dubbo.rpc.model.FrameworkModel;
-import org.apache.dubbo.rpc.model.PackableMethod;
-import org.apache.dubbo.rpc.protocol.tri.ClassLoadUtil;
-import org.apache.dubbo.rpc.protocol.tri.ExceptionUtils;
 import org.apache.dubbo.rpc.protocol.tri.RequestMetadata;
-import org.apache.dubbo.rpc.protocol.tri.TripleHeaderEnum;
-import org.apache.dubbo.rpc.protocol.tri.compressor.Compressor;
-import org.apache.dubbo.rpc.protocol.tri.compressor.Identity;
-import org.apache.dubbo.rpc.protocol.tri.observer.ClientCallToObserverAdapter;
-import org.apache.dubbo.rpc.protocol.tri.stream.ClientStream;
-import org.apache.dubbo.rpc.protocol.tri.stream.ClientStreamListener;
-import org.apache.dubbo.rpc.protocol.tri.stream.StreamUtils;
 
-import com.google.protobuf.Any;
-import com.google.rpc.DebugInfo;
-import com.google.rpc.ErrorInfo;
-import com.google.rpc.Status;
-
-import java.io.IOException;
-import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
-import java.util.concurrent.Executor;
 
-public class ClientCall {
+/**
+ * ClientCall does not care about transport layer details.
+ */
+public interface ClientCall {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(ClientCall.class);
-    private final Connection connection;
-    private final Executor executor;
-    private final FrameworkModel frameworkModel;
-    private RequestMetadata requestMetadata;
-    private ClientStream stream;
-    private ClientCall.Listener listener;
-    private boolean canceled;
-    private boolean headerSent;
-    private boolean autoRequestN = true;
+    /**
+     * Listener for receive response.
+     */
+    interface Listener {
 
-    public ClientCall(Connection connection, Executor executor, FrameworkModel frameworkModel) {
-        this.connection = connection;
-        this.executor = executor;
-        this.frameworkModel = frameworkModel;
-    }
-
-    public void sendMessage(Object message) {
-        if (canceled) {
-            throw new IllegalStateException("Call already canceled");
-        }
-        if (!headerSent) {
-            headerSent = true;
-            stream.sendHeader(requestMetadata.toHeaders());
-        }
-        final byte[] data;
-        try {
-            data = requestMetadata.packableMethod.packRequest(message);
-            int compressed =
-                Identity.MESSAGE_ENCODING.equals(requestMetadata.compressor.getMessageEncoding())
-                    ? 0 : 1;
-            final byte[] compress = requestMetadata.compressor.compress(data);
-            stream.writeMessage(compress, compressed);
-        } catch (Throwable t) {
-            LOGGER.error(String.format("Serialize triple request failed, service=%s method=%s",
-                requestMetadata.service,
-                requestMetadata.method), t);
-            cancel("Serialize request failed", t);
-            listener.onClose(TriRpcStatus.INTERNAL.withDescription("Serialize request failed")
-                .withCause(t), null);
-        }
-    }
-
-
-    public void requestN(int n) {
-        stream.requestN(n);
-    }
-
-    public void halfClose() {
-        if (!headerSent) {
-            return;
-        }
-        if (canceled) {
-            return;
-        }
-        stream.halfClose();
-    }
-
-    public void setCompression(String compression) {
-        this.requestMetadata.compressor = Compressor.getCompressor(frameworkModel, compression);
-    }
-
-    public StreamObserver<Object> start(RequestMetadata metadata,
-        ClientCall.Listener responseListener) {
-        this.requestMetadata = metadata;
-        this.listener = responseListener;
-        this.stream = new ClientStream(frameworkModel, executor, connection.getChannel(),
-            new ClientStreamListenerImpl(responseListener, metadata.packableMethod));
-        return new ClientCallToObserverAdapter<>(this);
-    }
-
-    public void cancel(String message, Throwable t) {
-        if (canceled) {
-            return;
-        }
-        // did not create stream
-        if (!headerSent) {
-            return;
-        }
-        canceled = true;
-        if (stream == null) {
-            return;
-        }
-        TriRpcStatus status = TriRpcStatus.CANCELLED.withCause(t);
-        if (message != null) {
-            status = status.withDescription(message);
-        } else {
-            status = status.withDescription("Cancel by client without message");
-        }
-        stream.cancelByLocal(status);
-    }
-
-    public boolean isAutoRequestN() {
-        return autoRequestN;
-    }
-
-    public void setAutoRequestN(boolean autoRequestN) {
-        this.autoRequestN = autoRequestN;
-    }
-
-    public interface Listener {
-
+        /**
+         * Called when the call is started, user can use this to set some configurations.
+         *
+         * @param call call implementation
+         */
         void onStart(ClientCall call);
 
+        /**
+         * Callback when message received.
+         *
+         * @param message message received
+         */
         void onMessage(Object message);
 
+        /**
+         * Callback when call is finished.
+         *
+         * @param status   response status
+         * @param trailers response trailers
+         */
         void onClose(TriRpcStatus status, Map<String, Object> trailers);
     }
 
-    class ClientStreamListenerImpl implements ClientStreamListener {
+    /**
+     * Send reset to server, no more data will be sent or received.
+     *
+     * @param t cause
+     */
+    void cancelByLocal(Throwable t);
 
-        private final Listener listener;
-        private final PackableMethod packableMethod;
-        private boolean done;
+    /**
+     * Request max n message from server
+     *
+     * @param messageNumber max message number
+     */
+    void request(int messageNumber);
 
-        ClientStreamListenerImpl(Listener listener, PackableMethod packableMethod) {
-            this.listener = listener;
-            this.packableMethod = packableMethod;
-        }
+    /**
+     * Send message to server
+     *
+     * @param message request to send
+     */
+    void sendMessage(Object message);
 
-        @Override
-        public void onStart() {
-            listener.onStart(ClientCall.this);
-        }
+    /**
+     * @param metadata         request metadata
+     * @param responseListener the listener to receive response
+     * @return the stream observer representing the request sink
+     */
+    StreamObserver<Object> start(RequestMetadata metadata,
+        Listener responseListener);
 
-        @Override
-        public void onMessage(byte[] message) {
-            if (done) {
-                LOGGER.warn(
-                    "Received message from closed stream,connection=" + connection + " service="
-                        + requestMetadata.service + " method="
-                        + requestMetadata.method.getMethodName());
-                return;
-            }
-            try {
-                final Object unpacked = packableMethod.parseResponse(message);
-                listener.onMessage(unpacked);
-            } catch (IOException | ClassNotFoundException e) {
-                cancelByErr(TriRpcStatus.INTERNAL.withDescription("Deserialize response failed")
-                    .withCause(e));
-            }
-        }
+    /**
+     * @return true if this call is auto request
+     */
+    boolean isAutoRequest();
 
-        @Override
-        public void complete(TriRpcStatus status, Map<String, Object> attachments,
-            Map<String, String> excludeHeaders) {
-            done = true;
-            final TriRpcStatus detailStatus;
-            final TriRpcStatus statusFromTrailers = getStatusFromTrailers(excludeHeaders);
-            if (statusFromTrailers != null) {
-                detailStatus = statusFromTrailers;
-            } else {
-                detailStatus = status;
-            }
-            try {
-                listener.onClose(detailStatus, StreamUtils.toAttachments(attachments));
-            } catch (Throwable t) {
-                cancelByErr(
-                    TriRpcStatus.INTERNAL.withDescription("Close stream error").withCause(t));
-            }
-        }
+    /**
+     * Set auto request for this call
+     *
+     * @param autoRequest whether auto request is enabled
+     */
+    void setAutoRequest(boolean autoRequest);
 
-        void cancelByErr(TriRpcStatus status) {
-            stream.cancelByLocal(status);
-        }
 
-        TriRpcStatus getStatusFromTrailers(Map<String, String> metadata) {
-            if (null == metadata) {
-                return null;
-            }
-            // second get status detail
-            if (!metadata.containsKey(TripleHeaderEnum.STATUS_DETAIL_KEY.getHeader())) {
-                return null;
-            }
-            final String raw = (metadata.remove(TripleHeaderEnum.STATUS_DETAIL_KEY.getHeader()));
-            byte[] statusDetailBin = StreamUtils.decodeASCIIByte(raw);
-            ClassLoader tccl = Thread.currentThread().getContextClassLoader();
-            try {
-                final Status statusDetail = Status.parseFrom(statusDetailBin);
-                List<Any> detailList = statusDetail.getDetailsList();
-                Map<Class<?>, Object> classObjectMap = tranFromStatusDetails(detailList);
+    /**
+     * No more data will be sent.
+     */
+    void halfClose();
 
-                // get common exception from DebugInfo
-                TriRpcStatus status = TriRpcStatus.fromCode(statusDetail.getCode())
-                    .withDescription(TriRpcStatus.decodeMessage(statusDetail.getMessage()));
-                DebugInfo debugInfo = (DebugInfo) classObjectMap.get(DebugInfo.class);
-                if (debugInfo != null) {
-                    String msg = ExceptionUtils.getStackFrameString(
-                        debugInfo.getStackEntriesList());
-                    status = status.appendDescription(msg);
-                }
-                return status;
-            } catch (IOException ioException) {
-                return null;
-            } finally {
-                ClassLoadUtil.switchContextLoader(tccl);
-            }
+    /**
+     * Set compression algorithm for request.
+     *
+     * @param compression compression algorithm
+     */
+    void setCompression(String compression);
 
-        }
-
-        private Map<Class<?>, Object> tranFromStatusDetails(List<Any> detailList) {
-            Map<Class<?>, Object> map = new HashMap<>();
-            try {
-                for (Any any : detailList) {
-                    if (any.is(ErrorInfo.class)) {
-                        ErrorInfo errorInfo = any.unpack(ErrorInfo.class);
-                        map.putIfAbsent(ErrorInfo.class, errorInfo);
-                    } else if (any.is(DebugInfo.class)) {
-                        DebugInfo debugInfo = any.unpack(DebugInfo.class);
-                        map.putIfAbsent(DebugInfo.class, debugInfo);
-                    }
-                    // support others type but now only support this
-                }
-            } catch (Throwable t) {
-                LOGGER.error("tran from grpc-status-details error", t);
-            }
-            return map;
-        }
-    }
 }
