+++ dubbo-rpc/dubbo-rpc-triple/src/main/java/org/apache/dubbo/rpc/protocol/tri/call/ObserverToClientCallListenerAdapter.java
@@ -34,8 +34,8 @@ public class ObserverToClientCallListenerAdapter implements ClientCall.Listener 
     @Override
     public void onMessage(Object message) {
         delegate.onNext(message);
-        if (call.isAutoRequestN()) {
-            call.requestN(1);
+        if (call.isAutoRequest()) {
+            call.request(1);
         }
     }
 
@@ -51,8 +51,8 @@ public class ObserverToClientCallListenerAdapter implements ClientCall.Listener 
     @Override
     public void onStart(ClientCall call) {
         this.call = call;
-        if (call.isAutoRequestN()) {
-            call.requestN(1);
+        if (call.isAutoRequest()) {
+            call.request(1);
         }
     }
 }
