+++ dubbo-rpc/dubbo-rpc-triple/src/main/java/org/apache/dubbo/rpc/protocol/tri/observer/ClientCallToObserverAdapter.java
@@ -32,7 +32,7 @@ public class ClientCallToObserverAdapter<T> extends CancelableStreamObserver<T> 
     }
 
     public boolean isAutoRequestEnabled() {
-        return call.isAutoRequestN();
+        return call.isAutoRequest();
     }
 
     @Override
@@ -46,7 +46,7 @@ public class ClientCallToObserverAdapter<T> extends CancelableStreamObserver<T> 
 
     @Override
     public void onError(Throwable throwable) {
-        call.cancel(null, throwable);
+        call.cancelByLocal(throwable);
         this.terminated = true;
     }
 
@@ -61,7 +61,7 @@ public class ClientCallToObserverAdapter<T> extends CancelableStreamObserver<T> 
 
     @Override
     public void cancel(Throwable throwable) {
-        call.cancel("Canceled by app ", throwable);
+        call.cancelByLocal(throwable);
         this.terminated = true;
     }
 
@@ -72,11 +72,11 @@ public class ClientCallToObserverAdapter<T> extends CancelableStreamObserver<T> 
 
     @Override
     public void request(int count) {
-        call.requestN(count);
+        call.request(count);
     }
 
     @Override
     public void disableAutoRequest() {
-        call.setAutoRequestN(false);
+        call.setAutoRequest(false);
     }
 }
