+++ dubbo-rpc/dubbo-rpc-triple/src/main/java/org/apache/dubbo/rpc/protocol/tri/observer/ServerCallToObserverAdapter.java
@@ -23,7 +23,7 @@ import org.apache.dubbo.rpc.CancellationContext;
 import org.apache.dubbo.rpc.TriRpcStatus;
 import org.apache.dubbo.rpc.protocol.tri.CancelableStreamObserver;
 import org.apache.dubbo.rpc.protocol.tri.ServerStreamObserver;
-import org.apache.dubbo.rpc.protocol.tri.call.ServerCall;
+import org.apache.dubbo.rpc.protocol.tri.call.AbstractServerCall;
 
 import java.util.Map;
 
@@ -32,11 +32,11 @@ public class ServerCallToObserverAdapter<T> extends CancelableStreamObserver<T> 
 
     private static final Logger LOGGER = LoggerFactory.getLogger(CancelableStreamObserver.class);
     public final CancellationContext cancellationContext;
-    private final ServerCall call;
+    private final AbstractServerCall call;
     private Map<String, Object> attachments;
     private boolean terminated = false;
 
-    public ServerCallToObserverAdapter(ServerCall call,
+    public ServerCallToObserverAdapter(AbstractServerCall call,
         CancellationContext cancellationContext) {
         this.call = call;
         this.cancellationContext = cancellationContext;
@@ -51,8 +51,8 @@ public class ServerCallToObserverAdapter<T> extends CancelableStreamObserver<T> 
         return terminated;
     }
 
-    private void setTerminated(boolean terminated) {
-        this.terminated = terminated;
+    private void setTerminated() {
+        this.terminated = true;
     }
 
     @Override
@@ -61,7 +61,7 @@ public class ServerCallToObserverAdapter<T> extends CancelableStreamObserver<T> 
             throw new IllegalStateException(
                 "Stream observer has been terminated, no more data is allowed");
         }
-        call.writeMessage(data);
+        call.sendMessage(data);
     }
 
     @Override
@@ -75,16 +75,12 @@ public class ServerCallToObserverAdapter<T> extends CancelableStreamObserver<T> 
             return;
         }
         call.close(status, attachments);
-        setTerminated(true);
+        setTerminated();
     }
 
     @Override
     public void onCompleted() {
-        if (isTerminated()) {
-            return;
-        }
-        call.close(TriRpcStatus.OK, attachments);
-        setTerminated(true);
+        onCompleted(TriRpcStatus.OK);
     }
 
     public void setResponseAttachments(Map<String, Object> attachments) {
@@ -97,7 +93,11 @@ public class ServerCallToObserverAdapter<T> extends CancelableStreamObserver<T> 
     }
 
     public void cancel(Throwable throwable) {
-        cancellationContext.cancel(throwable);
+        if (terminated) {
+            return;
+        }
+        setTerminated();
+        call.cancelByLocal(throwable);
     }
 
     public boolean isTimeout(long cost) {
@@ -111,6 +111,6 @@ public class ServerCallToObserverAdapter<T> extends CancelableStreamObserver<T> 
 
     @Override
     public void request(int count) {
-        call.requestN(count);
+        call.request(count);
     }
 }
