--- dubbo-rpc/dubbo-rpc-triple/src/main/java/org/apache/dubbo/rpc/protocol/tri/stream/ClientStreamListener.java
@@ -1,45 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.dubbo.rpc.protocol.tri.stream;
-
-import org.apache.dubbo.rpc.TriRpcStatus;
-
-import java.util.Collections;
-import java.util.Map;
-
-public interface ClientStreamListener extends StreamListener {
-
-    void onStart();
-
-    /**
-     * Indicate the stream is closed . Will be called exactly once
-     *
-     * @param rpcStatus
-     * @param attachments
-     */
-    void complete(TriRpcStatus rpcStatus, Map<String, Object> attachments,
-        Map<String, String> excludeHeaders);
-
-    default void complete(TriRpcStatus rpcStatus, Map<String, Object> attachments) {
-        complete(rpcStatus, attachments, Collections.emptyMap());
-    }
-
-    default void complete(TriRpcStatus rpcStatus) {
-        complete(rpcStatus, Collections.emptyMap(), Collections.emptyMap());
-    }
-}
