+++ dubbo-rpc/dubbo-rpc-triple/src/test/java/org/apache/dubbo/rpc/protocol/tri/service/TriHealthImplTest.java
@@ -22,6 +22,7 @@ import org.apache.dubbo.rpc.RpcException;
 
 import io.grpc.health.v1.HealthCheckRequest;
 import io.grpc.health.v1.HealthCheckResponse;
+import io.grpc.health.v1.HealthCheckResponse.ServingStatus;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
@@ -50,51 +51,64 @@ public class TriHealthImplTest {
     public void testWatch() throws Exception {
         TriHealthImpl triHealth = new TriHealthImpl();
 
-        HealthCheckRequest request = HealthCheckRequest.newBuilder().build();
+        HealthCheckRequest request = HealthCheckRequest.newBuilder()
+            .setService("testWatch")
+            .build();
+        triHealth.setStatus(request.getService(), ServingStatus.SERVING);
         StreamObserver<HealthCheckResponse> streamObserver = new MockStreamObserver();
 
+        RpcContext.removeCancellationContext();
         // test watch
         triHealth.watch(request, streamObserver);
-        MockStreamObserver mockStreamObserver = (MockStreamObserver) streamObserver;
         Assertions.assertNotNull(RpcContext.getCancellationContext().getListeners());
-        HashMap<String, IdentityHashMap<StreamObserver<HealthCheckResponse>, Boolean>> watches = getWatches(triHealth);
+        HashMap<String, IdentityHashMap<StreamObserver<HealthCheckResponse>, Boolean>> watches = getWatches(
+            triHealth);
         Assertions.assertTrue(watches.containsKey(request.getService()));
         Assertions.assertTrue(watches.get(request.getService()).containsKey(streamObserver));
         Assertions.assertTrue(watches.get(request.getService()).get(streamObserver));
+        MockStreamObserver mockStreamObserver = (MockStreamObserver) streamObserver;
         Assertions.assertEquals(mockStreamObserver.getCount(), 1);
-        Assertions.assertEquals(mockStreamObserver.getResponse().getStatus(), HealthCheckResponse.ServingStatus.SERVING);
+        Assertions.assertEquals(mockStreamObserver.getResponse().getStatus(),
+            HealthCheckResponse.ServingStatus.SERVING);
 
         // test setStatus
-        triHealth.setStatus(request.getService(), HealthCheckResponse.ServingStatus.SERVICE_UNKNOWN);
+        triHealth.setStatus(request.getService(),
+            HealthCheckResponse.ServingStatus.SERVICE_UNKNOWN);
         Assertions.assertEquals(mockStreamObserver.getCount(), 2);
-        Assertions.assertEquals(mockStreamObserver.getResponse().getStatus(), HealthCheckResponse.ServingStatus.SERVICE_UNKNOWN);
+        Assertions.assertEquals(mockStreamObserver.getResponse().getStatus(),
+            HealthCheckResponse.ServingStatus.SERVICE_UNKNOWN);
 
         triHealth.enterTerminalState();
         Assertions.assertEquals(mockStreamObserver.getCount(), 3);
-        Assertions.assertEquals(mockStreamObserver.getResponse().getStatus(), HealthCheckResponse.ServingStatus.NOT_SERVING);
+        Assertions.assertEquals(mockStreamObserver.getResponse().getStatus(),
+            HealthCheckResponse.ServingStatus.NOT_SERVING);
 
         // test clearStatus
         turnOffTerminal(triHealth);
         triHealth.clearStatus(request.getService());
         Assertions.assertEquals(mockStreamObserver.getCount(), 4);
-        Assertions.assertEquals(mockStreamObserver.getResponse().getStatus(), HealthCheckResponse.ServingStatus.SERVICE_UNKNOWN);
+        Assertions.assertEquals(mockStreamObserver.getResponse().getStatus(),
+            HealthCheckResponse.ServingStatus.SERVICE_UNKNOWN);
 
         // test listener
         RpcContext.getCancellationContext().close();
         Assertions.assertTrue(watches.isEmpty());
     }
 
-    private void turnOffTerminal(TriHealthImpl triHealth) throws NoSuchFieldException, IllegalAccessException {
+    private void turnOffTerminal(TriHealthImpl triHealth)
+        throws NoSuchFieldException, IllegalAccessException {
         Field terminalField = triHealth.getClass().getDeclaredField("terminal");
         terminalField.setAccessible(true);
         terminalField.set(triHealth, false);
     }
 
-    private HashMap<String, IdentityHashMap<StreamObserver<HealthCheckResponse>, Boolean>> getWatches(TriHealthImpl triHealth) throws Exception {
+    private HashMap<String, IdentityHashMap<StreamObserver<HealthCheckResponse>, Boolean>> getWatches(
+        TriHealthImpl triHealth) throws Exception {
         Field watchersField = triHealth.getClass().getDeclaredField("watchers");
         watchersField.setAccessible(true);
         HashMap<String, IdentityHashMap<StreamObserver<HealthCheckResponse>, Boolean>> watches =
-                (HashMap<String, IdentityHashMap<StreamObserver<HealthCheckResponse>, Boolean>>) watchersField.get(triHealth);
+            (HashMap<String, IdentityHashMap<StreamObserver<HealthCheckResponse>, Boolean>>) watchersField.get(
+                triHealth);
         return watches;
     }
